# specs/001-realtime-alerts/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Zabbix Monitor Alerts API
  version: 1.0.0
  description: API for real-time alerts, acknowledgments, filtering, and history.
servers:
  - url: http://localhost:5000
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/alerts:
    get:
      summary: Get alerts with optional filters
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [new, acknowledged, resolved]
        - in: query
          name: severity
          schema:
            type: integer
            minimum: 0
            maximum: 5
        - in: query
          name: host
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Alerts response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"
        "500":
          description: Server error
  /api/alerts/{alert_id}/acknowledge:
    post:
      summary: Acknowledge an alert
      parameters:
        - in: path
          name: alert_id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcknowledgeRequest"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcknowledgeResponse"
        "400":
          description: Invalid request
        "404":
          description: Alert not found
        "500":
          description: Server error
  /api/alerts/history:
    get:
      summary: Get historical alerts by date range
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date-time
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Historical alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"
        "400":
          description: Invalid date range
        "500":
          description: Server error
  /api/alerts/{alert_id}/history:
    get:
      summary: Get history timeline for a specific alert
      parameters:
        - in: path
          name: alert_id
          required: true
          schema:
            type: string
        - in: query
          name: date_from
          schema:
            type: string
            format: date-time
        - in: query
          name: date_to
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Alert history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertHistoryResponse"
        "404":
          description: Alert not found
        "500":
          description: Server error
components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [status]
    Alert:
      type: object
      properties:
        id:
          type: string
        zabbix_event_id:
          type: string
        zabbix_problem_id:
          type: string
          nullable: true
        host:
          type: string
        alert_name:
          type: string
        severity:
          type: integer
          minimum: 0
          maximum: 5
        status:
          type: string
          enum: [new, acknowledged, resolved]
        timestamp:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
          nullable: true
        last_updated_at:
          type: string
          format: date-time
      required: [id, zabbix_event_id, host, alert_name, severity, status, created_at]
    AlertHistory:
      type: object
      properties:
        id:
          type: string
        alert_id:
          type: string
        status_change_from:
          type: string
          nullable: true
        status_change_to:
          type: string
        changed_at:
          type: string
          format: date-time
        changed_by:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
      required: [id, alert_id, status_change_to, changed_at]
    AcknowledgeRequest:
      type: object
      properties:
        operator_name:
          type: string
        reason:
          type: string
    AcknowledgeResponse:
      type: object
      properties:
        success:
          type: boolean
        alert:
          $ref: "#/components/schemas/Alert"
        acknowledgment:
          type: object
      required: [success]
    AlertListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        pagination:
          type: object
          properties:
            skip:
              type: integer
            limit:
              type: integer
            total:
              type: integer
      required: [success, data]
    AlertHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/AlertHistory"
      required: [success, data]